splitting out fixtures into their own section, added rest of cli tests
