+++ tests/test_data.py
@@ -1,36 +1,5 @@
-import pytest
-import os
-
 from silvimetric.resources import Data
 from silvimetric.resources import Storage
-from silvimetric.resources import Bounds
-
-import conftest
-
-@pytest.fixture(scope='session')
-def no_cell_line_pc():
-    yield 84100
-
-@pytest.fixture(scope='session')
-def no_cell_line_path():
-    path = os.path.join(os.path.dirname(__file__), "data",
-            "test_data_2.copc.laz")
-    assert os.path.exists(path)
-    yield os.path.abspath(path)
-
-@pytest.fixture(scope='session')
-def no_cell_line_pipeline():
-    path = os.path.join(os.path.dirname(__file__), "data",
-            "test_data_2.json")
-    assert os.path.exists(path)
-    yield os.path.abspath(path)
-
-@pytest.fixture(scope='session')
-def autzen_filepath() -> str:
-    path = os.path.join(os.path.dirname(__file__), "data",
-            "autzen-small.copc.laz")
-    assert os.path.exists(path)
-    yield os.path.abspath(path)
 
 class Test_Data(object):
